+++ src/solvers/strings/string_constraint_generator_float.cpp
@@ -77,7 +77,7 @@ exprt constant_float(const double f, const ieee_float_spect &float_spec)
 {
   ieee_floatt fl(float_spec);
   if(float_spec == ieee_float_spect::single_precision())
-    fl.from_float(f);
+    fl.from_float(narrow_cast<float>(f));
   else if(float_spec == ieee_float_spect::double_precision())
     fl.from_double(f);
   else
